Improve Demo/GlobalConfiguration
