+++ app/src/main/java/me/jessyan/mvparms/demo/app/ActivityLifecycleCallbacksImpl.java
@@ -1,0 +1,83 @@
+package me.jessyan.mvparms.demo.app;
+
+import android.app.Activity;
+import android.app.Application;
+import android.os.Build;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.widget.TextView;
+
+import me.jessyan.mvparms.demo.R;
+import timber.log.Timber;
+
+/**
+ * ================================================
+ * Created by JessYan on 04/09/2017 17:14
+ * Contact with jess.yan.effort@gmail.com
+ * Follow me on https://github.com/JessYanCoding
+ * ================================================
+ */
+
+public class ActivityLifecycleCallbacksImpl implements Application.ActivityLifecycleCallbacks {
+
+    @Override
+    public void onActivityCreated(Activity activity, Bundle savedInstanceState) {
+        Timber.w(activity + " - onActivityCreated");
+    }
+
+    @Override
+    public void onActivityStarted(Activity activity) {
+        Timber.w(activity + " - onActivityStarted");
+        if (!activity.getIntent().getBooleanExtra("isInitToolbar", false)) {
+            //由于加强框架的兼容性,故将 setContentView 放到 onActivityCreated 之后,onActivityStarted 之前执行
+            //而 findViewById 必须在 Activity setContentView() 后才有效,所以将以下代码从之前的 onActivityCreated 中移动到 onActivityStarted 中执行
+            activity.getIntent().putExtra("isInitToolbar", true);
+            //这里全局给Activity设置toolbar和title,你想象力有多丰富,这里就有多强大,以前放到BaseActivity的操作都可以放到这里
+            if (activity.findViewById(R.id.toolbar) != null) {
+                if (activity instanceof AppCompatActivity) {
+                    ((AppCompatActivity) activity).setSupportActionBar((Toolbar) activity.findViewById(R.id.toolbar));
+                    ((AppCompatActivity) activity).getSupportActionBar().setDisplayShowTitleEnabled(false);
+                } else {
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                        activity.setActionBar((android.widget.Toolbar) activity.findViewById(R.id.toolbar));
+                        activity.getActionBar().setDisplayShowTitleEnabled(false);
+                    }
+                }
+            }
+            if (activity.findViewById(R.id.toolbar_title) != null) {
+                ((TextView) activity.findViewById(R.id.toolbar_title)).setText(activity.getTitle());
+            }
+            if (activity.findViewById(R.id.toolbar_back) != null) {
+                activity.findViewById(R.id.toolbar_back).setOnClickListener(v -> {
+                    activity.onBackPressed();
+                });
+            }
+        }
+    }
+
+    @Override
+    public void onActivityResumed(Activity activity) {
+        Timber.w(activity + " - onActivityResumed");
+    }
+
+    @Override
+    public void onActivityPaused(Activity activity) {
+        Timber.w(activity + " - onActivityPaused");
+    }
+
+    @Override
+    public void onActivityStopped(Activity activity) {
+        Timber.w(activity + " - onActivityStopped");
+    }
+
+    @Override
+    public void onActivitySaveInstanceState(Activity activity, Bundle outState) {
+        Timber.w(activity + " - onActivitySaveInstanceState");
+    }
+
+    @Override
+    public void onActivityDestroyed(Activity activity) {
+        Timber.w(activity + " - onActivityDestroyed");
+    }
+}
